vb2_sha: Use uint64_t for total size in SHA-2
Message for SHA-2 padding is ending with the length of message in bits
written as 64-bit big-endian integer. Use uint64_t to store and
calculate the total size of message, so messages with length greater
than 512MiB can be hashed.

BRANCH=main
BUG=b:434894535, b:434895132
TEST=Use fastboot oem sha256 on large partition

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/7090335

+++ firmware/2lib/include/2sha.h
@@ -59,7 +59,7 @@ struct vb2_sha1_context {
 
 struct vb2_sha256_context {
 	uint32_t h[8];
-	uint32_t total_size;
+	uint64_t total_size;
 	uint32_t size;
 	uint8_t block[2 * VB2_SHA256_BLOCK_SIZE];
 };
@@ -70,7 +70,7 @@ struct vb2_sha256_context {
 
 struct vb2_sha512_context {
 	uint64_t h[8];
-	uint32_t total_size;
+	uint64_t total_size;
 	uint32_t size;
 	uint8_t block[2 * VB2_SHA512_BLOCK_SIZE];
 };
